Refactor claims to consolidate 'asset' parameter
I've had to remove ObservableFunctions.daml for the time being.  This is
the module providing an implementation for Claim (->) t a The reason is
that now the first parameter is of kind * -> * -> * -> * (instead of the
previous * -> * -> * that satisfies (->)).

I'd like to maintain back-compat but not sure how; that needs more
thought and I'll leave it for another day.