Rewrite `show claim` to reflect serialized type.
This is required because otherwise `And`s would show up as
`And (a) (b) (c) ...`, whereas we should explicitly show that the
last element is a list.
Although the above representation is neater, it may result in confusion
when debugging.
In other words, the only thing that our custom show instance should do
is drop the field labels, and add brackets, to make things more legible.