Remove startDate from input of `lifecycle`
This prevents an API break. The acquisition date is assumed to be undefined (minus infinity) until the first When node is crossed.

As a drawback, we cannot write contracts where the outermost node is a Scale (which would be a well-defined contract if the corresponding observable is a numerical constant)